implement add order to cart

+++ ku_market_place/models/order_item.py
@@ -3,10 +3,9 @@ from .product import Product
 
 
 class OrderItem(models.Model):
-
     product = models.ForeignKey(Product, on_delete=models.CASCADE)
     quantity = models.IntegerField(default=0)
-    total_amount = models.IntegerField(default=0)
+    total_amount = models.FloatField(default=0)
 
     def __str__(self):
         return self.product.productDisplayName
